# packages/protocol/openapi.yaml
openapi: 3.1.0
info:
  title: ClickClack API
  version: 0.1.0
  license:
    name: MIT
servers:
  - url: http://localhost:8080
paths:
...
  /api/workspaces/{workspace_id}:
    get:
      operationId: getWorkspace
      parameters:
        - $ref: "#/components/parameters/workspace_id"
      responses:
        "200":
          description: Workspace
...
  /api/channels/{channel_id}:
    patch:
      operationId: updateChannel
      parameters:
        - $ref: "#/components/parameters/channel_id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateChannelRequest"
      responses:
        "200":
          description: Updated channel
...
  /api/messages/{message_id}:
    get:
      operationId: getMessage
      parameters:
        - $ref: "#/components/parameters/message_id"
      responses:
        "200":
          description: Message visible to the current user
    patch:
      operationId: updateMessage
      parameters:
        - $ref: "#/components/parameters/message_id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMessageRequest"
      responses:
        "200":
          description: Updated message
    delete:
      operationId: deleteMessage
      parameters:
        - $ref: "#/components/parameters/message_id"
      responses:
        "200":
          description: Soft-deleted message
...
  /api/realtime/ephemeral:
    post:
      operationId: publishEphemeral
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EphemeralEventRequest"
      responses:
        "202":
          description: Ephemeral event accepted
...
components:
  parameters:
    workspace_id:
      name: workspace_id
      in: path
      required: true
      schema:
        type: string
    channel_id:
      name: channel_id
      in: path
      required: true
      schema:
        type: string
    message_id:
      name: message_id
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    UpdateChannelRequest:
      type: object
      properties:
        name:
          type: string
        kind:
          type: string
        archived:
          type: boolean
    CreateMessageRequest:
      type: object
      required: [body]
      properties:
        body:
          type: string
        body_format:
          type: string
          enum: [markdown]
          default: markdown
        quoted_message_id:
          type: string
          description: |
            Optional ID of a message to quote. The quoted message must live in
            the same channel, DM conversation, or thread as the new message;
            cross-context quoting is rejected with 400. The server captures a
            soft snapshot of the quoted body at send time.
        nonce:
          type: string
          maxLength: 128
          description: |
            Optional client idempotency key for retry-safe message creation.
            Reusing the same nonce with the same body and quote returns the
            existing message with HTTP 200 instead of creating a duplicate.
...
    EphemeralEventRequest:
      type: object
      required: [workspace_id, type]
      properties:
        workspace_id:
          type: string
        channel_id:
          type: string
        direct_conversation_id:
          type: string
        type:
          type: string
          enum: [typing.started, typing.stopped, presence.changed]
        payload:
          type: object
          additionalProperties: true
...